requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
brotlipy==0.7.0
cchardet==2.1.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
orjson==3.8.0
Pillow==9.2.0
py-cord==2.1.3
pycares==4.2.2
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
yarl==1.8.1
youtube-dl==2021.12.17
